[#21] Added initials docs.
