+++ docs/api/README.md
@@ -4,7 +4,8 @@ title: Introduction
 
 # API Reference
 
-
 The Druxt.js Schema module exports the following:
 - [DruxtSchemaModule Nuxt.js module](module)
+- [DruxtSchema class](schema)
 - [DruxtSchemaMixin Vue.js mixin](mixins/schema)
+- [druxtSchema Vuex store](stores/schema)
