+++ docs/guide/README.md
@@ -4,11 +4,11 @@ title: Introduction
 
 # Druxt.js Schema
 
-The Druxt.js Schema
-
 The Druxt.js Schema module uses the [Drupal](https://drupal.org) JSON:API and **Display mode** configuration to build a [Vuex](https://vuex.vuejs.org) based Schema store to power your Drupal driven [Nuxt.js](https://nuxtjs.org) frontend.
 
 
 ## How it works?
 
-...
+Upon installation the module installs a Nuxt.js plugin, Vuex store and builds the required Entity Form and View schemas.
+
+Schemas are loaded into the Vuex store on-demand through the usage of the provided plugin.
